chore: refactor genlocal into genbot (#3674)
Previously if you wanted to run genlocal you needed to have a
bunch of dependencies installed. Some of which can be a little
tricky to install correctly. These dependencies happen to be the
same as genbot. For this reason this commit refactors genlocal
into genbot so that the genbot in local mode can be run from
a docker container that has the exact same deps installed as our
nightly regen.